requirements.txt
# mkl-fft==1.3.6
# mkl-random @ file:///Users/ec2-user/mkl/mkl_random_1682994911338/work
# mkl-service==2.4.0
# numpy @ file:///Users/ec2-user/mkl/numpy_and_numpy_base_1682973000331/work
# Pillow==9.3.0
# torch==1.7.1
# torchvision==0.8.2
# typing_extensions @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_cazhnzmri4/croot/typing_extensions_1681939520610/work


flask
Bs4
Pillow==9.3.0
requests
pybase64
git+https://github.com/openai/CLIP.git

